+++ pam/internal/gdm/extension_test.go
